correct delay     correct delay

+++ utilities/reverse_api/run.py
@@ -46,7 +46,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='kitty farm')
     parser.add_argument('number', type=str, help='folder catgirls-n')
     parser.add_argument('-d', '--delay', type=float, default=0, help='Delay time in seconds (default is 0)')
-    parser.add_argument('-m', '--max', type=float, default=10, help='Maximum number of failed redirects before killing process (default is 10)')
+    parser.add_argument('-m', '--max', type=int, default=10, help='Maximum number of failed redirects before killing process (default is 10)')
     parser.add_argument('-t', '--test', type=bool, default=False, help='Runs the program with a testing cookie file named test_cookies.json (default is False)')
     args = parser.parse_args()
